requirements.txt
anyio==4.2.0
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==4.1.5
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-storages==1.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
filelock==3.14.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gunicorn==20.1.0
idna==3.7
keyboard==0.13.5
MouseInfo==0.1.3
mysqlclient==2.2.1
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
PyAutoIt==0.6.5
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
rsa==4.9
setuptools==70.0.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.2
virtualenv==20.26.2
wheel==0.43.0
